requirements.txt
discord.py
dotenv
aiohttp
google-genai
flask